Duplicate command/unless for pam-update-auth
The problem is that there is no "check" command for the pam-auth-update
command. For now, the command will just get run every time puppet runs
as an `unless` parameter, and if that fails, it will run again as
`command` in order to elevate the error message to the puppet report.

It's a hack. Maybe someday somebody will find a better way to do it.